telegram-parser/requirements.txt
# Telegram Parser Dependencies
telethon>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0